+++ src/main/java/compiler/nodes/statement_nodes/NodePrint.java
@@ -2,8 +2,6 @@ package compiler.nodes.statement_nodes;
 
 
 import compiler.Generator;
-import compiler.nodes.expression_nodes.term_nodes.ArrayAccess;
-import compiler.nodes.expression_nodes.term_nodes.IdentExpression;
 import compiler.nodes.expression_nodes.term_nodes.NodeTerm;
 import compiler.nodes.expression_nodes.term_nodes.StringExpression;
 
@@ -34,11 +32,11 @@ public class NodePrint implements NodeStatement {
 
     private void printNumArr(Generator generator, String name) {
         String size = "sizeof(" + name + ") / sizeof(" + name +  "[0])";
-        String condition = "i <" + size;
+        String condition = "__lc__ <" + size;
         generator.appendContents("printf(\"[\");\n");
-        generator.appendContents("for (int i = 0; " + condition + "; i++) {\n");
-        generator.appendContents("printf(\"%d\", " + name + "[i]);\n");
-        generator.appendContents("if (i + 1 !=" + size + ") printf(\", \");\n");
+        generator.appendContents("for (int __lc__ = 0; " + condition + "; __lc__++) {\n");
+        generator.appendContents("printf(\"%d\", " + name + "[__lc__]);\n");
+        generator.appendContents("if (__lc__ + 1 !=" + size + ") printf(\", \");\n");
         generator.appendContents("}\n");
         generator.appendContents("printf(\"]\\n\");\n");
     }
@@ -47,12 +45,12 @@ public class NodePrint implements NodeStatement {
 
         String top = "sizeof(" + name + ")";
         String size = "sizeof(" + name + ") / sizeof(" + name +  "[0])";
-        String condition = "i <" + size;
+        String condition = "__lc__ <" + size;
         generator.appendContents("printf(\"[\");\n");
         generator.appendContents("if (" + top + "!= 0 ) {");
-        generator.appendContents("for (int i = 0; " + condition + "; i++) {\n");
-        generator.appendContents("printf(\"%s\", " + name + "[i]);\n");
-        generator.appendContents("if (i + 1 !=" + size + ") printf(\", \");\n");
+        generator.appendContents("for (int __lc__ = 0; " + condition + "; __lc__++) {\n");
+        generator.appendContents("printf(\"%s\", " + name + "[__lc__]);\n");
+        generator.appendContents("if (__lc__ + 1 !=" + size + ") printf(\", \");\n");
         generator.appendContents("}\n");
         generator.appendContents("}\n");
         generator.appendContents("printf(\"]\\n\");\n");
@@ -62,57 +60,40 @@ public class NodePrint implements NodeStatement {
 
         String top = "sizeof(" + name + ")";
         String size = "sizeof(" + name + ") / sizeof(" + name +  "[0])";
-        String condition = "i <" + size;
+        String condition = "__lc__ <" + size;
         generator.appendContents("printf(\"[\");\n");
         generator.appendContents("if (" + top + "!= 0 ) {");
-        generator.appendContents("for (int i = 0; " + condition + "; i++) {\n");
-        generator.appendContents("printf(\"%c\", " + name + "[i]);\n");
-        generator.appendContents("if (i + 1 !=" + size + ") printf(\", \");\n");
+        generator.appendContents("for (int __lc__ = 0; " + condition + "; __lc__++) {\n");
+        generator.appendContents("printf(\"%c\", " + name + "[__lc__]);\n");
+        generator.appendContents("if (__lc__ + 1 !=" + size + ") printf(\", \");\n");
         generator.appendContents("}\n");
         generator.appendContents("}\n");
         generator.appendContents("printf(\"]\\n\");\n");
     }
 
     public void operator(Generator generator) {
-                                                        
-        
+               
         if (term instanceof StringExpression) {
             generator.appendContents("    printf(" + term.toString() + ");\n");
             return;
         }
-        
-        if (term instanceof IdentExpression || term instanceof ArrayAccess) {
-            String name = term.getToken().getValue();
-            switch (returnType) {
-                case "numeric":
-                case "array|numeric":
-                    if (returnType.equals("array|numeric") && (term instanceof IdentExpression)) {
-                        printNumArr(generator, name);
-                        return;
-                    }
-                    generator.appendContents("printf(\"%d\\n\", ");
-                    break;
-                case "string":
-                case "array|string":
-                    if (returnType.equals("array|string") && (term instanceof IdentExpression)) {
-                        printStringArr(generator, name);
-                        return;
-                    }
-                    generator.appendContents("printf(\"%s\\n\", ");
-                    break;
-                case "char":
-                case "array|char":
-                    if (returnType.equals("array|char") && (term instanceof IdentExpression)) {
-                        printCharArr(generator, name);
-                        return;
-                    }
-                    generator.appendContents("printf(\"%c\\n\", ");
-                    break;
-                default:
-            }
-        } else {
-
-            generator.appendContents("printf(\"%d\\n\", ");
+        System.out.println("returnType is: " + returnType);
+        switch (returnType) {
+            case "string":
+                generator.appendContents("printf(\"%s\\n\", ");
+                break;
+            case "numeric":
+                generator.appendContents("printf(\"%d\\n\", ");
+                break;
+            case "char":
+                generator.appendContents("printf(\"%c\\n\", ");
+                break;
+            case "array|numeric":
+                printNumArr(generator, term.getToken().getValue()); return;
+            case "array|string":
+                printStringArr(generator, term.getToken().getValue()); return;
+            case "array|char":
+                printCharArr(generator, term.getToken().getValue()); return;
         }
         term.operator(generator);
         generator.appendContents(");");