+++ src/main/java/compiler/nodes/statement_nodes/loops/NodeLoop.java
@@ -7,13 +7,23 @@ import compiler.nodes.statement_nodes.NodeStatement;
 public class NodeLoop implements NodeStatement {
     
     private NodeScope scope = null;
+    private String count;
 
     public NodeLoop(NodeScope scope) {
         this.scope = scope;
     }
 
+    public NodeLoop(NodeScope scope, String count) {
+        this.scope = scope;
+        this.count = count;
+    }
+
     public NodeScope getScope() {
         return this.scope;
+    }
+
+    public String getCount() {
+        return this.count;
     }
 
     @Override
@@ -25,7 +35,10 @@ public class NodeLoop implements NodeStatement {
     }
 
     public void operator(Generator generator) {
-        generator.appendContents("    while (1)\n");
+        if (count == null)
+            generator.appendContents("    while (1)\n");
+        else
+            generator.appendContents("for (int __lc__ = 0; __lc__ < " + count + "; __lc__++)\n");
         scope.operator(generator);
     }